+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This `CHANGELOG.md` implements the spirit of http://keepachangelog.com/.
 
-## [1.30](https://github.com/Comcast/eel/compare/v1.29...dev) - [Unreleased]
+## [1.31](https://github.com/Comcast/eel/compare/v1.30.0...dev) - [Unreleased]
+
+## [1.30](https://github.com/Comcast/eel/compare/v1.29...v1.30.0) - 2019-06-07
+* XRULES-13840: Default handlers in EEL
+* XRULES-13983: Flow execution log issues
 
 ## [1.29](https://github.com/Comcast/eel/compare/v1.28...v1.29) - 2019-05-17